No one saw that

+++ Source/Voxel/SkyfallenPillars.h
@@ -21,5 +21,5 @@ public:
 	FastNoise perlin;
 	FastNoise cellular;
 
-	int folliageDensity;
+	double folliageDensity;
 };
